Fix: use URI and IPAddress API to validate services IP
This fix originates from an issue encountered with gRPC call that
continues to take a lot of resource if an invalid address like "x.x.x."
was passed. The issue didn't occcur with well formatted invalid address.
The downside is that this fix only allows IP v4 IP addresses as valid
inputs.